[+] Fixing FindBugs report, iteration 2

+++ libs/kiwi/src/main/java/com/hyperrealm/kiwi/io/Base64InputStream.java
@@ -116,7 +116,7 @@ public class Base64InputStream extends FilterInputStream {
             c = -1;
         }
 
-        return ((int) r);
+        return r;
     }
 
     /**
